APL-VH-Android: November 2023 Release of APL 2023.3 compliant Viewhost Code

+++ apl/src/main/java/com/amazon/apl/android/TextMeasure.java
@@ -155,8 +155,8 @@ public class TextMeasure {
 
         // return results
         final float[] measurement = new float[]{
-                mMetricsTransform.toCore(mMeasuredWidthPx),
-                mMetricsTransform.toCore(mMeasuredHeightPx)};
+                mMetricsTransform.toCore((float)mMeasuredWidthPx),
+                mMetricsTransform.toCore((float)mMeasuredHeightPx)};
 
         return measurement;
     }
